--- a/packages/api/package.json
+++ b/packages/api/package.json
@@ -43,8 +43,8 @@
     "json-schema-traverse": "^1.0.0",
     "lodash.merge": "^4.6.2",
     "make-dir": "^3.1.0",
-    "memoizee": "^0.4.15",
-    "oas": "^18.1.0",
+    "oas": "^18.2.0",
+    "object-hash": "^3.0.0",
     "ts-morph": "^14.0.0"
   },
   "devDependencies": {
@@ -54,8 +54,8 @@
     "@types/find-cache-dir": "^3.2.1",
     "@types/js-yaml": "^4.0.5",
     "@types/lodash.merge": "^4.6.6",
-    "@types/memoizee": "^0.4.7",
     "@types/mocha": "^9.1.0",
+    "@types/object-hash": "^2.2.1",
     "chai": "^4.3.6",
     "eslint": "^8.10.0",
     "mocha": "^9.2.1",