… packages/api/test/__fixtures__/sdk/readme/index.ts