| 🚥 Fix RM-3929 |
| :-- |

## 🧰 Changes

This refactors how we're generating TS types during codegen to never generate TS code for the same schema more than once. This work doesn't make a huge dent in the still slow codegen process, but it's an improvement for huge specs.

#### Before

```
  typescript generator
    ✔ should generate typescript code (3137ms)
    ✔ should be able to generate valid TS when a body is optional but metadata isnt (2056ms)
    ✔ should work against the petstore (17054ms)
    ✔ should work against our OAS (78918ms)
```

#### After

```
  typescript generator
    ✔ should generate typescript code (4143ms)
    ✔ should be able to generate valid TS when a body is optional but metadata isnt (2323ms)
    ✔ should work against the petstore (16065ms)
    ✔ should work against our OAS (66688ms)
```